Added binary output file by plant for data visualization
The LP.OUT file is a binary file with all of the LP variables stored for
each of the systems solved.  This will allow another program to be
created that will read the binary file and allow a plant by plant
visualization for each period, water year and outage profile.

I also cleaned up a few format strings for the output file directory.