Fixed variable initialization and formatting
Trap2 Verification and Enhancement
December 20, 2018
Copied the current official Trap2 program and components and recompiled and linked to verify that results could be exactly duplicated.
Issue:
Part of output was garbage with asterisks filling spaces where data should have been.
Solution:
Found 3 variables that needed to be initialized to zero. Apparently the compiler used for the official version either automatically initialized all variables or a switch was set to do that. Variables were initialized and output matched the official output exactly.
Issue:
Text label on output file wraps around and has to be manually edited for the Trap post processor to work.
Solution:
Used explicit format (instead of an *) to remove the wrap around.
Action: December 20, 2018
The modified version was pushed back into the GitHub repository.